backend/requirements.txt
# Web framework
fastapi==0.95.2
uvicorn[standard]==0.22.0

# Database
sqlalchemy==1.4.41
psycopg2-binary==2.9.6
alembic==1.11.1

# API
pydantic==1.10.8
python-dotenv==1.0.0
newsapi-python==0.2.7
requests==2.31.0

# Testing
pytest==7.3.1
pytest-cov==4.1.0

# Utils
python-dateutil==2.8.2